-Add Button

+++ Views/ToDoListView.swift
@@ -19,19 +19,26 @@ struct ToDoListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ToDoListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsListView()
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ToDoListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform:listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                 }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform:listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-         }
-        .listStyle(PlainListStyle())
+            
+        }
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
         leading: EditButton(),
